boilerplate/mcp-backend/foundry-lab/requirements.txt
openai>=1.0.0
azure-identity>=1.15.0
azure-ai-projects>=1.0.0b3
azure-search-documents>=11.4.0
python-dotenv>=1.0.0
rich>=13.0.0
